Move `order_*` and `median_of_3*` functions to parent module, add `median` convenience function.
