+++ src/select/qselect.rs
@@ -1,5 +1,4 @@
 use std::cmp::Ordering;
-use std::mem;
 
 // The threshold below which qselect_inplace_by() should just sort the slice.
 const SORT_THRESH: usize = 6;
@@ -83,7 +82,7 @@ fn partition_by<T, F: FnMut(&T, &T) -> Ordering>(data: &mut [T], mut ord_fn: F) 
         let mid = (&data[len / 2], len / 2);
         let last = (&data[len - 1], len - 1);
 
-        median_of_3_by(&first, &mid, &last, |left, right| ord_fn(left.0, right.0)).1
+        super::median_of_3_by(&first, &mid, &last, |left, right| ord_fn(left.0, right.0)).1
     };
 
     data.swap(pivot_idx, len - 1);
@@ -102,36 +101,6 @@ fn partition_by<T, F: FnMut(&T, &T) -> Ordering>(data: &mut [T], mut ord_fn: F) 
     curr
 }
 
-/// Of the three values passed, return the median.
-pub fn median_of_3<T: Ord>(x: T, y: T, z: T) -> T {
-    median_of_3_by(x, y, z, <T as Ord>::cmp)
-}
-
-/// Given an ordering function, of the three values passed, return the median.
-pub fn median_of_3_by<T, F: FnMut(&T, &T) -> Ordering>(mut x: T, mut y: T, mut z: T, mut ord_fn: F) -> T {
-    in_order_by(&mut x, &mut y, &mut ord_fn);
-    in_order_by(&mut x, &mut z, &mut ord_fn);
-    in_order_by(&mut y, &mut z, &mut ord_fn);
-
-    y
-}
-
-/// If `x > y`, swap `x` and `y`.
-#[inline]
-pub fn in_order<T: Ord>(x: &mut T, y: &mut T) {
-    if x > y {
-        mem::swap(x, y);
-    }
-}
-
-/// Given an ordering function, if `x > y`, swap `x` and `y`.
-#[inline]
-pub fn in_order_by<T, F: FnMut(&T, &T) -> Ordering>(x: &mut T, y: &mut T, mut ord_fn: F) {
-    if ord_fn(x, y) == Ordering::Greater {
-        mem::swap(x, y);
-    }
-}
-
 #[test]
 fn test_qsel() {
     macro_rules! test_qsel (
@@ -145,24 +114,4 @@ fn test_qsel() {
     test_qsel!(1, 2, 3, 4, 5; 3; 4);
     test_qsel!(5, 4, 3, 2, 1; 1; 2);
     test_qsel!(2, 2, 3, 3, 4; 1; 2);
-}
-
-#[test]
-fn test_median() {
-    macro_rules! test_median {
-        ($x:expr, $y:expr, $z:expr; $expected:expr) => (
-            let x = $x;
-            let y = $y;
-            let z = $z;
-            let result = median_of_3(&x, &y, &z);
-            assert_eq!(*result, $expected);
-        )
-    }
-    
-    test_median!(1, 2, 3; 2);
-    test_median!(1, 3, 2; 2);
-    test_median!(2, 1, 3; 2);
-    test_median!(2, 3, 1; 2);
-    test_median!(3, 1, 2; 2);
-    test_median!(3, 2, 1; 2);
 }
